Move processor in-flight tracker
