+++ v2/processor.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"sync"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
@@ -248,54 +247,6 @@ func (p *Processor) waitForInFlightHandlers() error {
 		}
 		return fmt.Errorf("timed out waiting %s for %d in-flight message handler(s): %w", *gracePeriod, p.inFlight.count(), context.DeadlineExceeded)
 	}
-}
-
-type inFlightHandlers struct {
-	mu      sync.Mutex
-	active  int
-	drained chan struct{}
-}
-
-func newInFlightHandlers() *inFlightHandlers {
-	drained := make(chan struct{})
-	close(drained)
-	return &inFlightHandlers{
-		drained: drained,
-	}
-}
-
-func (h *inFlightHandlers) add() {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-
-	if h.active == 0 {
-		h.drained = make(chan struct{})
-	}
-	h.active++
-}
-
-func (h *inFlightHandlers) done() {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-
-	h.active--
-	if h.active == 0 {
-		close(h.drained)
-	}
-}
-
-func (h *inFlightHandlers) doneCh() <-chan struct{} {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-
-	return h.drained
-}
-
-func (h *inFlightHandlers) count() int {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-
-	return h.active
 }
 
 type PanicHandlerOptions struct {
